# backend/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AI-Library-Backend-API",
    "description": "Backend API Document"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "paths": {
...
    "/api/books/search": {
      "get": {
        "tags": ["Books"],
        "summary": "搜索书籍",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "搜索结果"
          }
        }
      }
    },
    "/api/books/popular/list": {
      "get": {
        "tags": ["Books"],
        "summary": "获取热门书籍",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "热门书籍列表"
          }
        }
      }
    },
...
    "/api/books": {
      "post": {
        "tags": ["Books"],
        "summary": "创建书籍",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "book_id": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string"
                  },
                  "author": {
                    "type": "string"
                  },
                  "publisher": {
                    "type": "string"
                  },
                  "publication_year": {
                    "type": "integer"
                  },
                  "call_no": {
                    "type": "string"
                  },
                  "language": {
                    "type": "string"
                  },
                  "doc_type": {
                    "type": "string"
                  }
                },
                "example": {
                  "book_id": "TEST_BOOK_001",
                  "title": "Test Book",
                  "author": "Test Author",
                  "publisher": "Test Publisher",
                  "publication_year": 2023,
                  "call_no": "TP312/TEST",
                  "language": "English",
                  "doc_type": "Computer Science"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功"
          }
        }
      }
    },
    "/api/books/{id}": {
      "get": {
        "tags": ["Books"],
        "summary": "获取书籍详情",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "书籍详情"
          }
        }
      },
      "put": {
        "tags": ["Books"],
        "summary": "更新书籍",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "author": {
                    "type": "string"
                  },
                  "publisher": {
                    "type": "string"
                  },
                  "publication_year": {
                    "type": "integer"
                  },
                  "call_no": {
                    "type": "string"
                  },
                  "language": {
                    "type": "string"
                  },
                  "doc_type": {
                    "type": "string"
                  }
                },
                "example": {
                  "title": "Test Book Updated",
                  "author": "Test Author",
                  "publisher": "Test Publisher",
                  "publication_year": 2024,
                  "call_no": "TP312/TEST",
                  "language": "English",
                  "doc_type": "Computer Science"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功"
          }
        }
      },
      "delete": {
        "tags": ["Books"],
        "summary": "删除书籍",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          }
        }
      }
    },
    "/api/books/{id}/related": {
      "get": {
        "tags": ["Books"],
        "summary": "获取相关书籍",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "相关书籍列表"
          }
        }
      }
    },
    "/api/books/all": {
      "get": {
        "tags": ["Books"],
        "summary": "获取所有书籍",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "书籍列表"
          }
        }
      }
    },
    "/api/books/batch": {
      "post": {
        "tags": ["Books"],
        "summary": "批量删除书籍",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "ids": ["BOOK_001", "BOOK_002"]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "批量删除成功"
          }
        }
      }
    },
...
  }
}